requirements.txt
fastapi
uvicorn
pysentimiento
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
feedparser
spacy